refactor(commands): one module per subcommand
commands.lua held the argument tree, the completion, and nine
handlers in 958 lines. The handlers are what changes: ten of the
last ten commits to the file touched one.

  commands/init.lua              286  the :CodeDiff tree, completion, entry
  commands/parse.lua             112  argument handling the handlers share
  commands/handlers/explorer.lua 162  :CodeDiff
  commands/handlers/git_diff.lua 140  :CodeDiff [rev] [rev2]
  commands/handlers/merge.lua    104  :CodeDiff merge <file>
  commands/handlers/history.lua   97  :CodeDiff history
  .../explorer_staged.lua         77  :CodeDiff --staged
  .../file_diff.lua               56  :CodeDiff file <a> <b>
  .../dir_diff.lua                51  :CodeDiff dir <a> <b>

Every function moved was renamed. Each handler is now `M.run`
in the module named after its subcommand, so `handle_explorer`
is `handlers.explorer.run` -- the name is the path. The two
merge-base wrappers compute a base and then delegate, so each
is `run_merge_base` alongside the handler it delegates to,
which leaves no calls between handler modules at all. The four
shared argument helpers keep their names under `parse.`.

Bodies moved unchanged. The only edits were the renames above
and prefixing calls that crossed a new module boundary.

Nothing outside changed. Every caller reaches for
`require("codediff.commands")`, which init.lua answers.

Behaviour compared before and after across seven subcommands --
HEAD, file A...B, file a b, explorer, --staged, history, dir --
recording layout, merge flag, single_side, window count, panel,
both buffers' contents and the diff size: byte-identical.
